pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "skyview-cli"
version = "0.1.0"
description = "Browse astronomical sky images by RA/Dec — CLI + Jupyter"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
dependencies = [
    "click>=8.0",
    "requests>=2.28",
    "matplotlib>=3.6",
    "astropy>=5.0",
    "astroquery>=0.4",
    "Pillow>=9.0",
    "numpy>=1.22",
]

[project.optional-dependencies]
jupyter = ["ipywidgets>=8.0", "IPython>=8.0"]

[project.scripts]
skyview = "skyview.cli:main"

[tool.setuptools.packages.find]
where = ["src"]
